load_json_files/requirements.txt
Django==2.2.10
django-crispy-forms==1.7.2
pyresparser
gunicorn
django-allauth
django-clear-cache
djangorestframework